refactor(package): modify package
Lotto 도메인의 패키지를 분리했습니다.
