+++ src/main/java/lotto/domain/lotto/LottoCreator.java
@@ -1,8 +1,9 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.List;
 
 public class LottoCreator {
+
     public Lotto createLotto(List<Integer> lottoNumbers) {
         return new Lotto(lottoNumbers.stream()
                 .map(LottoNumber::new)
