+++ src/main/java/lotto/domain/lotto/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 public interface LottoGenerator {
     Lotto generateLotto();
