+++ src/test/java/lotto/domain/LottoTest.java
@@ -7,6 +7,8 @@ import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import lotto.common.exception.DuplicateLottoNumberException;
 import lotto.common.exception.InvalidLottoSizeException;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
